+++ src/components/CustomBottomText.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { StyleSheet, Text, View } from 'react-native'
 
-import { COLORS } from '../utils'
 import { CustomBottomTextProps } from '../types/props'
+import { COLORS } from '../utils'
 
 const classes = StyleSheet.create({
   container:{