+++ src/components/GoBack.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
+import { StyleSheet, TouchableOpacity } from 'react-native'
 import { Icon } from 'react-native-elements'
 
 import CustomBelowButtonText from './CustomBelowButtonText'
+import { GoBackProps } from '../types/props'
 import { COLORS } from '../utils'
 
 const classes = StyleSheet.create({
@@ -11,21 +12,19 @@ const classes = StyleSheet.create({
   }
 })
 
-const GoBack = () => {
-  return (
-    <View style={classes.container}>
-      <Icon
-        iconStyle={{
-          marginTop: 0.75
-        }}
-        color={COLORS.RED}
-        name='chevron-left'
-        type='material'
-        size={16}
-      />
-      <CustomBelowButtonText text='Go back' />
-    </View>
-  )
-}
+const GoBack = ({ onPress }: GoBackProps) => (
+  <TouchableOpacity style={classes.container} onPress={onPress}>
+    <Icon
+      iconStyle={{
+        marginTop: 0.75
+      }}
+      color={COLORS.RED}
+      name='chevron-left'
+      type='material'
+      size={16}
+    />
+    <CustomBelowButtonText text='Go back' />
+  </TouchableOpacity>
+)
 
 export default GoBack