+++ changelog.md
@@ -1,7 +1,13 @@
 # Frontend App
 
+## Version 0.4.0
+  - Implemented:
+    - UI of the first screen of the [design](https://www.figma.com/file/qDfkoP6JJnu4f19ZmrW5Xo/IHC-Project?node-id=0%3A1).
+    - Fixed some style bugs in the CustomButton.
+    - Navigation across the first and second screen.
+
 ## Version 0.3.0
-  - Implemented
+  - Implemented:
     - UI of the second screen of the [design](https://www.figma.com/file/qDfkoP6JJnu4f19ZmrW5Xo/IHC-Project?node-id=0%3A1).
     - Many reusable components to use across the app.
     - Previous configuration for navigation.